product and json files merged
